Moved fetchmode.h to includes
