+++ src/ace/managers/viewport/simplebuffer.c
@@ -7,7 +7,7 @@
 #include <ace/utils/tag.h>
 #include <ace/utils/extview.h>
 #include <ace/generic/screen.h> // Has the look up table for the COPPER_X_WAIT values.
-#include "fetchmode.h"
+#include <ace/managers/viewport/fetchmode.h>
 #ifdef AMIGA
